use numpy.float64

+++ io_handler.py
@@ -18,11 +18,11 @@ def read_data(filename="solar_jfc.dat"):
         vel_z = []
         for i, row in enumerate(reader):
             if i > 0:
-                pos_x.append(float(row[2]) * au)
-                pos_y.append(float(row[3]) * au)
-                pos_z.append(float(row[4]) * au)
-                vel_x.append(float(row[5]) * au / day)
-                vel_y.append(float(row[6]) * au / day)
-                vel_z.append(float(row[7]) * au / day)
-                masses.append(float(row[1]) * m_sol)
+                pos_x.append(np.float64(row[2]) * au)
+                pos_y.append(np.float64(row[3]) * au)
+                pos_z.append(np.float64(row[4]) * au)
+                vel_x.append(np.float64(row[5]) * au / day)
+                vel_y.append(np.float64(row[6]) * au / day)
+                vel_z.append(np.float64(row[7]) * au / day)
+                masses.append(np.float64(row[1]) * m_sol)
     return masses, pos_x, pos_y, pos_z, vel_x, vel_y, vel_z
